+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- sort queried events (#417)
+
 ## [0.2.0] - 2021-11-02
 
 ### Changed